[DragValue] Use rounded values for checking if value needs setting
If the old value is not rounded, the modified if-expression would always
trigger clearing the edit memory, and hence making editing by keyboard
impossible -- the edit buffer would only last a single frame.

The old_value could also be always rounded if we rounded the result of
parsing the keyboard input, but this prevents the user from inputting
values more precise than allowed by default, which I think is a useful
behavior worth keeping.